@idreyn @nathan-weinberg comments
