+++ server/background.py
@@ -5,8 +5,7 @@ import schedule
 
 # Copied from https://schedule.readthedocs.io/en/stable/background-execution.html
 
-
-def run_continuously(interval=1):
+def run_continuously(interval_s=1) -> threading.Event:
     """Continuously run, while executing pending jobs at each
     elapsed time interval.
     @return cease_continuous_run: threading. Event which can
@@ -24,8 +23,12 @@ def run_continuously(interval=1):
         def run(cls):
             while not cease_continuous_run.is_set():
                 schedule.run_pending()
-                time.sleep(interval)
+                time.sleep(interval_s)
 
-    continuous_thread = ScheduleThread()
+    # preston added daemon=True. By making it a daemon thread, the main thread getting a signal
+    # will cause only the background thread to remain, and Python will then exit fully. Without that,
+    # the background thread will continue to run.
+    # See https://docs.python.org/3/library/threading.html#threading.Thread.daemon
+    continuous_thread = ScheduleThread(daemon=True)
     continuous_thread.start()
     return cease_continuous_run
